Use GetUserId for update & dalete

+++ api/Controllers/AccountController.cs
@@ -26,11 +26,16 @@ public class AccountController(IAccountRepository accountRepository) : BaseApiCo
 
         return loggedInDto;
     }
-    
+
     [Authorize]
-    [HttpDelete("delete/{userId}")]
+    [HttpDelete("delete")]
     public async Task<ActionResult<DeleteResult>> DeleteById(string userId, CancellationToken cancellationToken)
     {
+        var memberId = User.GetUserId();
+
+        if (memberId is null)
+            return Unauthorized("You are not logged. Please log in again.");
+
         DeleteResult? deleteResult = await accountRepository.DeleteByIdAsync(userId, cancellationToken);
 
         if (deleteResult is null)
